abstracted mongo_transaction in _base_worker process pre_existing task now do the same jobs as the real time tasks
abstracted _base_dd_insert as helper func not in any class

defined new errorType of data_ref_not_exist

isolated exit_code_class
error logging implemented around data unplickling

+++ src/worker.py
@@ -104,17 +104,17 @@ class worker(base_worker):
         self.event_cnt+=1
         event_cnt = self.event_cnt
         global process_id
-        try:
-            assert j["_id"]['_data'] not in process_id
-            process_id.add( j["_id"]['_data'])
-            self.client['log']['log'].insert_one( { 'packetID' : j["_id"]['_data'],
-                                                   'activity'  : 'threadStarted' } )
-        except KeyboardInterrupt:
-            raise
-        except AssertionError:
-            logging_info = j["_id"] + ' already in process_id'
-            print(logging_info)
-            self.logger.info(logging_info)
+        # try:   removed due to mongo db provide good ID already
+        #     assert j["_id"]['_data'] not in process_id
+        #     process_id.add( j["_id"]['_data'])
+        #     self.client['log']['log'].insert_one( { 'packetID' : j["_id"]['_data'],
+        #                                            'activity'  : 'threadStarted' } )
+        # except KeyboardInterrupt:
+        #     raise
+        # except AssertionError:
+        #     logging_info = j["_id"] + ' already in process_id'
+        #     print(logging_info)
+        #     self.logger.info(logging_info)
         
         logging_info = 'processed' + str(event_cnt) + 'packets'
         print(logging_info)
@@ -125,21 +125,39 @@ class worker(base_worker):
         data_unpickled = pickle.loads(doc ['data'])
         print('data_tyoe '+ str(type(data_unpickled)) + ' received')
         # print('data type : '  , type(data_unpickled) )
-        if  type(data_unpickled) is data_ref.data_ref:
+        try:
+            if  type(data_unpickled) is data_ref.data_ref:
+                
+                data = data_unpickled.deref_data(mongoClient = self.client)
+                self.logger.info('deref_data')
+                # print('deref_data')
+            elif  issubclass(type(data_unpickled),  worker_command.worker_command):
+                self.process_common_command(command = data_unpickled)
+                return
+                pass
+            else :
+                # assume data_unpickled is direct data
+                data = data_unpickled
+                self.logger.info('raw_data')
+            import process_data
+        
+            process_data.process_data(data)
+        except Exception as e:
             
-            data = data_unpickled.deref_data(mongoClient = self.client)
-            self.logger.info('deref_data')
-            # print('deref_data')
-        elif  issubclass(type(data_unpickled),  worker_command.worker_command):
-            self.process_common_command(command = data_unpickled)
-            return
-            pass
-        else :
-            # assume data_unpickled is direct data
-            data = data_unpickled
-            self.logger.info('raw_data')
-        import process_data
-        process_data.process_data(data)
+            from data_ref import _base_dd_insert
+            err_data_pickled, err_data_ref = _base_dd_insert(e, 'error', 'py_Exception_Store', 'error', 'py_Exception_Ref' ,mongoClient = self.client)
+            from copy import deepcopy
+            err_doc = deepcopy(doc)
+            err_doc['data_ref_error_doc_ID'] = err_data_ref.documentID
+            from _base_worker import mongo_transaction
+            db_from = worker_db
+            col_from = self.name
+            db_to = 'error'
+            col_to = 'error_job'
+            logging_info = 'error found' + e.__repr__()
+            mongo_transaction(err_doc, self.client, db_from, col_from, db_to, col_to, logging_info)
+            
+            
 
     def check_document_integrity(self, doc):
         if 'tag' in doc:
@@ -169,32 +187,74 @@ class worker(base_worker):
             x.start()
         else:
             self.process_event_threadable(j)
+    def update_doc_start_process(self, doc):
+        myid = doc["_id"]
+        self.client[worker_db][self.worker_collection_name].update_one(
+                {
+                    "_id" : myid, 
+                },
+                {"$set" : {'worker_start_time' : datetime.now()}}
+                
+            )
+    def update_doc_end_process(self, doc):
+        myid = doc["_id"]
+        self.client[worker_db][self.worker_collection_name].update_one(
+                {
+                    "_id" : myid, 
+                },
+                {"$set" : {'worker_end_time' : datetime.now()}}
+                
+            )
     def process_event_threadable(self, j):   # worker thread process data
         # process existing doc
         
         doc = j['fullDocument']
         if not self.check_document_integrity(doc):
             return
-
+        self.update_doc_start_process(doc)
         self.process_doc(doc)
+        self.update_doc_end_process(doc)
         self.record_post_threadable_event(j)        # worker level log 
     def record_post_threadable_event(self, j):
+        """
+        make the logs after task finish,
+        log time and event description
+        """
         self.logging_doc_results(j)   # doc level logging
         try:
-            insert_result = self.client[worker_db]['availableWorker'].insert_one({'_id' : self.worker_collection_name,
-                                                             'free-since' : int(time.time())})
+            insert_result = self.client[worker_db]['availableWorker'].insert_one(
+                {'_id' : self.worker_collection_name,
+                 'free-since' : int(time.time())})
         except DuplicateKeyError:
             pass
         logging_info = self.worker_collection_name + ' is now free'
         print(logging_info)
         self.logger.info(logging_info)
         doc = j['fullDocument']
-        self.client[worker_db][self.worker_collection_name].delete_one(doc)
-        logging_info = 'packet with _id', j["_id"]['_data'] + 'processed and removed from worker'
+        
+        log_packet_ref_on = True
+        if log_packet_ref_on:
+            
+            from _base_worker import processed_data_col, mongo_transaction
+            mongo_client = self.client
+            db_from = worker_db
+            db_to   = "eventTrigger"
+            col_to = processed_data_col
+            col_from = self.worker_collection_name
+            collection_from = mongo_client[db_from][col_from]
+            collection_to= mongo_client[db_to][col_to]
+            logging_info = 'packet id' + str(doc['_id']) + ' finished and logged to ' + col_to +" " +col_to
+            mongo_transaction(doc , mongo_client, db_from, col_from, db_to, col_to, logging_info)
+            
+        else:
+            self.client[worker_db][self.worker_collection_name].delete_one(doc)
+        logging_info = ('packet with _id', str(j['fullDocument']["_id"]) 
+                        + 'processed and removed from worker')
         print(logging_info)
         self.logger.info(logging_info)
-        self.client['log']['log'].insert_one( { 'packetID' : j["_id"]['_data'],
-                                                       'activity'  : 'data_processed' } )
+        self.client['log']['log'].insert_one( 
+            { 'packetID' : str(j['fullDocument']["_id"]) ,
+             'activity'  : 'data_processed' } )
         pass
         
 
@@ -204,6 +264,12 @@ class worker(base_worker):
 
     
     def pre_work(self):
+        """
+        initialisations before worker start working
+        related to worker and control implementation specifically for eachg worker
+
+        """
+        
         self.event_cnt = 0
         
         self.work_listenStream = self.eventStream
